+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/Chip.cs
@@ -14,7 +14,6 @@ namespace System.Device.Gpio.Libgpiod.V2;
 /// allows callers to retrieve information about each line, watch lines for state changes and make line requests.
 /// </summary>
 /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__chips.html"/>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal class Chip : LibGpiodProxyBase
 {
     private readonly int _chipNumber;
