+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/EdgeEvent.cs
@@ -14,7 +14,6 @@ namespace System.Device.Gpio.Libgpiod.V2;
 /// of events are being read.
 /// </summary>
 /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__edge__event.html"/>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal class EdgeEvent : LibGpiodProxyBase
 {
     private readonly EdgeEventSafeHandle _handle;
