+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/LibGpiodChipInfo.cs
@@ -12,7 +12,6 @@ namespace System.Device.Gpio.Libgpiod.V2;
 /// The chip info contains all the publicly available information about a chip.
 /// </summary>
 /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__chip__info.html"/>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal class LibGpiodChipInfo : LibGpiodProxyBase
 {
     private readonly int _chipNumber;
