+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/LibGpiodProxyBase.cs
@@ -7,7 +7,6 @@ using System.Runtime.InteropServices;
 
 namespace System.Device.Gpio.Libgpiod.V2;
 
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal abstract class LibGpiodProxyBase : IDisposable
 {
     /// <remarks>
