+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/LibGpiodProxyFactory.cs
@@ -8,7 +8,6 @@ using LibgpiodV2 = Interop.LibgpiodV2;
 
 namespace System.Device.Gpio.Libgpiod.V2;
 
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal static class LibGpiodProxyFactory
 {
     /// <summary>
