+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/LineConfig.cs
@@ -21,7 +21,6 @@ namespace System.Device.Gpio.Libgpiod.V2;
 /// If any of the offsets was duplicated, the last one will take precedence.
 /// </summary>
 /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__line__config.html"/>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal class LineConfig : LibGpiodProxyBase
 {
     internal LineConfigSafeHandle Handle { get; }
