+++ src/System.Device.Gpio/Interop/Unix/libgpiod/V2/Proxies/LineInfo.cs
@@ -14,7 +14,6 @@ namespace System.Device.Gpio.Libgpiod.V2;
 /// The line must be requested to access the line value.
 /// </summary>
 /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__line__info.html"/>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 internal class LineInfo : LibGpiodProxyBase
 {
     private readonly LineInfoSafeHandle _handle;
