+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/LibGpiodV2Driver.cs
@@ -17,13 +17,13 @@ using static Interop;
 namespace System.Device.Gpio.Drivers;
 
 /// <summary>
-/// Driver that uses libgpiod V2 for GPIO control.
+/// Driver that uses libgpiod V2 ABI for GPIO control.
 /// <remarks>
-/// At the time of this writing, this driver is only available when compiling from source. See instructions at
-/// https://libgpiod.readthedocs.io/en/latest/building.html.
+/// The versioning of the driver libraries is confusing. This one uses the V2 ABI, which is implemented by the V3 libgpiod library,
+/// so this one typically works when libgpiod.so.3 is available in the path.
+/// The driver has replaced the previous versions on all Raspberry PI OS versions based on Debian 13 (trixie), released Dec 25.
 /// </remarks>
 /// </summary>
-[Experimental(DiagnosticIds.SDGPIO0001, UrlFormat = DiagnosticIds.UrlFormat)]
 public sealed class LibGpiodV2Driver : UnixDriver
 {
     private static readonly string ConsumerId = $"C#-{nameof(LibGpiodV2Driver)}-{Process.GetCurrentProcess().Id}";
@@ -89,6 +89,17 @@ public sealed class LibGpiodV2Driver : UnixDriver
         }
 
         return ret;
+    }
+
+    /// <summary>
+    /// Get the API version of the library as a human-readable string
+    /// </summary>
+    /// <seealso href="https://libgpiod.readthedocs.io/en/latest/group__misc.html#gac7919e728ad7c9ba534ec543a8dbcac2"/>
+    /// <exception cref="GpiodException">Unexpected error invoking native function</exception>
+    public static string GetApiVersion()
+    {
+        return LibGpiodProxyBase.CallLibgpiod(() =>
+            Marshal.PtrToStringAuto(LibgpiodV2.gpiod_api_version()) ?? throw new GpiodException($"Could not get API version: {LastErr.GetMsg()}"));
     }
 
     /// <inheritdoc/>
